GRANITE-46835: fix bugs (#9)
Because of some incompatibility issues we decided to maintain Java 8 in this project. However from January 31, 2024 Java 8 will reach end of support in Google App Engine. So if we want to make changes after that we have to update this project to Java 11+ which we should do anyway, since the newer versions of Oak is compiled with Java 11.  

---------

Co-authored-by: Steffen Van <svan@adobe.com>